+++ test/integration/braintree/customer_session_gateway_spec.js
@@ -6,10 +6,11 @@ const {
   CreateCustomerSessionInput,
   CustomerRecommendationsInput,
   CustomerSessionInput,
+  MonetaryAmountInput,
+  PayPalPurchaseUnitInput,
   PhoneInput,
   UpdateCustomerSessionInput,
   RecommendedPaymentOption,
-  Recommendations,
 } = require("./../../../lib/braintree/graphql");
 
 describe("CustomerSessionGateway", () => {
@@ -51,6 +52,12 @@ describe("CustomerSessionGateway", () => {
     return gateway.customerSession.createCustomerSession(input);
   }
 
+  function buildPurchaseUnit() {
+    const amount = new MonetaryAmountInput("10.00", "USD");
+
+    return PayPalPurchaseUnitInput.builder(amount).build();
+  }
+
   describe("Create customer session", () => {
     it("create a customer session without email and phone", async () => {
       const input = CreateCustomerSessionInput.builder()
@@ -76,6 +83,16 @@ describe("CustomerSessionGateway", () => {
 
     it("create a customer session with API derived session id", async () => {
       const input = CreateCustomerSessionInput.builder().build();
+      const result = await gateway.customerSession.createCustomerSession(input);
+
+      assert.isTrue(result.success);
+      assert.isDefined(result.target);
+    });
+
+    it("create a customer session with purchase units", async () => {
+      const input = CreateCustomerSessionInput.builder()
+        .purchaseUnits([buildPurchaseUnit()])
+        .build();
       const result = await gateway.customerSession.createCustomerSession(input);
 
       assert.isTrue(result.success);
@@ -112,6 +129,7 @@ describe("CustomerSessionGateway", () => {
       );
       const input = UpdateCustomerSessionInput.builder(sessionId)
         .customer(customer)
+        .purchaseUnits([buildPurchaseUnit()])
         .build();
 
       const result = await gateway.customerSession.updateCustomerSession(input);
@@ -140,16 +158,26 @@ describe("CustomerSessionGateway", () => {
 
   describe("Get customer recommendations", () => {
     it("gets customer recommendations", async () => {
-      const customer = buildCustomerSessionInput(
-        "PR5_test@example.com",
-        "4085005005"
-      );
-      const customerRecommendationsInput = CustomerRecommendationsInput.builder(
-        "11EF-A1E7-A5F5EE5C-A2E5-AFD2801469FC",
-        [Recommendations.PAYMENT_RECOMMENDATIONS]
-      )
-        .customer(customer)
+      const customer = CustomerSessionInput.builder()
+        .hashedEmail(
+          "48ddb93f0b30c475423fe177832912c5bcdce3cc72872f8051627967ef278e08"
+        )
+        .hashedPhoneNumber(
+          "a2df2987b2a3384210d3aa1c9fb8b627ebdae1f5a9097766c19ca30ec4360176"
+        )
+        .deviceFingerprintId("00DD010662DE")
+        .userAgent(
+          "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/x.x.x.x Safari/537.36"
+        )
         .build();
+
+      const customerRecommendationsInput =
+        CustomerRecommendationsInput.builder()
+          .sessionId("94f0b2db-5323-4d86-add3-paypal000000")
+          .customer(customer)
+          .purchaseUnits([buildPurchaseUnit()])
+          .domain("domain.com")
+          .build();
 
       const result = await gateway.customerSession.getCustomerRecommendations(
         customerRecommendationsInput
@@ -167,29 +195,46 @@ describe("CustomerSessionGateway", () => {
         RecommendedPaymentOption.PAYPAL
       );
       assert.equal(recommendation.recommendedPriority, 1);
+
+      const paymentRecommendation =
+        payload.recommendations.paymentRecommendations[0];
+
+      assert.equal(
+        paymentRecommendation.paymentOption,
+        RecommendedPaymentOption.PAYPAL
+      );
+      assert.equal(paymentRecommendation.recommendedPriority, 1);
     });
 
-    it("does not get recommendations for non-existent session", async () => {
-      const customer = buildCustomerSessionInput(
-        "PR9_test@example.com",
-        "4085005009"
-      );
-
-      const customerRecommendationsInput = CustomerRecommendationsInput.builder(
-        "11EF-34BC-2702904B-9026-C3ECF4BAC765",
-        [Recommendations.PAYMENT_RECOMMENDATIONS]
-      )
-        .customer(customer)
+    it("does not get recommendations for unauthorized session", (done) => {
+      const customer = CustomerSessionInput.builder()
+        .deviceFingerprintId("00DD010662DE")
+        .userAgent(
+          "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/x.x.x.x Safari/537.36"
+        )
         .build();
 
-      const result = await gateway.customerSession.getCustomerRecommendations(
-        customerRecommendationsInput
-      );
+      const customerRecommendationsInput =
+        CustomerRecommendationsInput.builder()
+          .sessionId("94f0b2db-5323-4d86-add3-paypal000000")
+          .customer(customer)
+          .purchaseUnits([buildPurchaseUnit()])
+          .domain("domain.com")
+          .merchantAccountId("gbp_pwpp_multi_account_merchant_account")
+          .build();
 
-      assert.isFalse(result.success);
-      assert.ok(
-        result.errors.deepErrors()[0].message.includes("does not exist")
-      );
+      gateway.customerSession
+        .getCustomerRecommendations(customerRecommendationsInput)
+        .then(() => {
+          done(assert.fail("Expected authorization error."));
+        })
+        .catch((err) => {
+          if (err.type === "authorizationError") {
+            done();
+          } else {
+            done(assert.fail("Expected authorization error."));
+          }
+        });
     });
   });
 });