+++ test/unit/braintree/graphql/customer_recommendations_payload_spec.js
@@ -4,17 +4,17 @@ const {
   CustomerRecommendations,
   CustomerRecommendationsPayload,
   RecommendedPaymentOption,
-  PaymentOptions,
+  PaymentRecommendation,
 } = require("../../../../lib/braintree/graphql");
 
 describe("CustomerRecommendationsPayload", () => {
   it("should correctly return isInPayPalNetwork and recommendations", () => {
-    const paymentOptions = new PaymentOptions(
+    const paymentRecommendations = new PaymentRecommendation(
       RecommendedPaymentOption.PAYPAL,
       1
     );
     const customerRecommendations = new CustomerRecommendations([
-      paymentOptions,
+      paymentRecommendations,
     ]);
     const payload = new CustomerRecommendationsPayload(
       true,